# [#64] No leading zeros