--- a/benchmark/benchmark.toml
+++ b/benchmark/benchmark.toml
@@ -6,6 +6,7 @@ cash = true
 words = [ "all", 'strings', """are the same""", '''type''']
 bool = [false, true, false]
 today = 1979-05-27T07:32:00Z  # ZonedTime
+ints = [0, 1, 2, 120, 9000, -5, +12, -100]
 # digits = [1.0, 2.0, nan, -nan, +inf, -inf, 1.0]  # not for TOML v0.4.0
 # date = [1895-05-27, 1980-11-23, 2018-10-25]  # [Day] not for TOML v0.4.0