--- a/src/Toml/Parser/Core.hs
+++ b/src/Toml/Parser/Core.hs
@@ -13,12 +13,10 @@ import Control.Applicative (Alternative (empty))
 import Data.Text (Text)
 import Data.Void (Void)
 
-import Text.Megaparsec (Parsec, anySingle, errorBundlePretty, match, parse,
-                        satisfy, try, (<?>), eof, oneOf, lookAhead)
+import Text.Megaparsec (Parsec, anySingle, errorBundlePretty, match, parse, satisfy, try, (<?>))
 import Text.Megaparsec.Char (alphaNumChar, char, digitChar, eol, hexDigitChar, space, space1,
                              string, tab)
-import Text.Megaparsec.Char.Lexer (binary, float, hexadecimal, octal, signed, skipLineComment,
-                                   symbol, decimal)
+import Text.Megaparsec.Char.Lexer (binary, float, hexadecimal, octal, signed, skipLineComment, symbol)
 import qualified Text.Megaparsec.Char.Lexer as L (lexeme, space)