--- a/src/Toml/Parser/Value.hs
+++ b/src/Toml/Parser/Value.hs
@@ -12,22 +12,22 @@ module Toml.Parser.Value
        ) where
 
 import Control.Applicative (Alternative (some, (<|>)))
-import Control.Applicative.Combinators (between, count, manyTill, option, optional, sepEndBy,
-                                        skipMany, sepBy1)
+import Control.Applicative.Combinators (between, count, manyTill, option, optional, sepBy1,
+                                        sepEndBy, skipMany)
 
 import Data.Char (chr, isControl)
-import Text.Read (readMaybe)
 import Data.Fixed (Pico)
 import Data.Semigroup ((<>))
 import Data.Text (Text)
 import Data.Time (LocalTime (..), ZonedTime (..), fromGregorianValid, makeTimeOfDayValid,
                   minutesToTimeZone)
+import Text.Read (readMaybe)
 
 import Toml.Parser.Core (Parser, alphaNumChar, anySingle, binary, char, digitChar, eol, float,
-                         hexDigitChar, hexadecimal, lexeme, octal, satisfy, sc, signed,
-                         space, string, tab, text, try, (<?>))
+                         hexDigitChar, hexadecimal, lexeme, octal, satisfy, sc, signed, space,
+                         string, tab, text, try, (<?>))
 import Toml.PrefixTree (Key (..), Piece (..))
-import Toml.Type (DateTime (..), UValue (..), AnyValue, typeCheck)
+import Toml.Type (AnyValue, DateTime (..), UValue (..), typeCheck)
 
 import qualified Control.Applicative.Combinators.NonEmpty as NC
 import qualified Data.Text as Text
@@ -275,13 +275,12 @@ arrayP = lexeme (between (char '[' *> sc) (char ']') elements) <?> "array"
 
 
 valueP :: Parser UValue
-valueP = UBool    <$> boolP
+valueP = UText    <$> textP
+     <|> UBool    <$> boolP
+     <|> UArray   <$> arrayP
      <|> UDate    <$> dateTimeP
      <|> UDouble  <$> try doubleP
      <|> UInteger <$> integerP
-     <|> UText    <$> textP
-     <|> UArray   <$> arrayP
-
 
 anyValueP :: Parser AnyValue
 anyValueP = typeCheck <$> valueP >>= \case