Addressing #64 

I re-implemented `decimalP`, it's much simpler now. 
I've uncommented the test on leading zeroes, and modified the `does not parse numbers` tests to actually test failure and not success. To make those pass I had to use the `lookAhead` trick.

While testing these, I realized that the string `"a = 123.3b=23"` is accepted and parsed like `"a = 123.3\nb=23"`. Same for `"a = trueb=23"` and others. It's kind of a weird case, I don't know if it's worth spending time on. Note: `"a = 123b=23"` does not parse because I've used the `lookAhead` trick for integers, so it's definitely fixable.