@ChShersh 2018-11-07
> and modified the does not parse numbers tests to actually test failure and not success

This is kinda weird how those tests passed before 🤔 

> While testing these, I realized that the string "a = 123.3b=23" is accepted and parsed

@jiegillet Is this the old behavior before your change or after?