@jiegillet 2018-11-07
> This is kinda weird how those tests passed before 🤔

Well, they were weird tests... Example of old tests:
```haskell
it "does not parse numbers prefixed with `0B`" $ parseInteger "0B1101" 0
```
`parseInteger` would consume the leading `0`, stop there return 0 and pass the test, which makes sense, but if you actually want the parser to fail then the test should really be `integerFailOn "0B1101"`. But then you need to modify the parser so that it only stops consuming digits when it hits a space or something.
 
> > While testing these, I realized that the string "a = 123.3b=23" is accepted and parsed
> 
> @jiegillet Is this the old behavior before your change or after?

That's technically both because `123.3` is a float and I only modified the parser for integers. Before the change "a=1b=2" is a valid toml file but after not.  "a=1 b=2" is valid though, and that maybe should be another discussion whether it should or not.