@jiegillet 2018-11-09
@willbasky, you wrote the original tests, what's your take on the question?