@jiegillet 2018-11-13
I've decided to remove the part about `lookAhead`, I will open a separate issue about it, because it's more general than being about leading zeros. 
I've also reverted the changes to the tests, except the ones relevant to leading zeros.
This is now a very straightforward change to the integer parser.