@jiegillet 2018-11-27
I did another proper benchmark after my last commit.
Current master:
```
time                 287.4 μs   (274.3 μs .. 303.4 μs)
                     0.989 R²   (0.984 R² .. 0.996 R²)
mean                 291.0 μs   (286.4 μs .. 295.1 μs)
std dev              14.58 μs   (11.97 μs .. 18.00 μs)
variance introduced by outliers: 29% (moderately inflated)
```

This PR:
```
benchmarked Parse/tomland
time                 282.7 μs   (270.1 μs .. 294.3 μs)
                     0.992 R²   (0.990 R² .. 0.997 R²)
mean                 283.5 μs   (279.0 μs .. 288.8 μs)
std dev              15.61 μs   (11.67 μs .. 21.23 μs)
variance introduced by outliers: 32% (moderately inflated)
```

It's a 3% speedup, with similar moderately inflated variances. 3% speedup doesn't mean much, but at least I'm not slowing things down anymore.